+++ CHANGES.rst
@@ -4,6 +4,10 @@ Changelog
 1.2 (unreleased)
 ----------------
 
+- Add plone.abovecontenttitle viewlet manager to person, organization, position
+  and contact (held_position) views.
+  [vincentfretin]
+
 - js functions have a namespace.
   [thomasdesvenain]