+++ CHANGES.rst
@@ -12,6 +12,9 @@ Changelog
   a main position, all current positions, closed positions.
   [thomasdesvenain]
 
+- Sort get_held_positions on organization.
+  [cedricmessiant]
+
 - Add plone.abovecontenttitle viewlet manager to person, organization, position
   and contact (held_position) views.
   [vincentfretin]